added openldap, sudo cookbooks and updated ssh and authorized_keys cookbooks to use new dependencies

+++ cookbooks/authorized_keys/files/default/manage_authorized_keys.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 if [ $# == 0 ]; then
-  echo "Usage: $0 <gid_1> <gid_2>"
+  echo "Usage: $0 <group_1> <group_2>"
   exit 1
 fi
 
@@ -10,22 +10,23 @@ declare -A LDAP_USERS
 #
 # Setup authorized_keys for users in groups assigned to this server
 #
-for GID in "$@"
+for GROUP in "$@"
 do
-  USERS=$(ldapsearch -LLL -x -H ldap://ldap1.bos1 -b dc=bos1 "gidNumber=${GID}" uid | grep "uid: " | cut -f2 -d' ')
+  USERS=$(getent group $GROUP | cut -d':' -f4 | tr -s ',' '\n')
 
   for USER in $USERS
   do
     HOME_DIR="/home/$USER"
 
-#    if [ ! -s "$HOME_DIR/.ssh/authorized_keys" ]; then
+    if [ ! -s "$HOME_DIR/.ssh/authorized_keys" ]; then
+      echo "Adding authorized_keys file for ${USER}"
       mkdir -p $HOME_DIR/.ssh
       wget -O $HOME_DIR/.ssh/authorized_keys https://raw.github.com/basho/ssh_keys/master/$USER
       chmod 700 $HOME_DIR
       chmod 700 $HOME_DIR/.ssh
       chmod 600 $HOME_DIR/.ssh/authorized_keys
-      chown -R $USER:$GID $HOME_DIR
-#    fi
+      chown -R $USER:$GROUP $HOME_DIR
+    fi
 
     LDAP_USERS[$USER]="1"
   done
@@ -38,6 +39,7 @@ LOCAL_USERS=$(ls /home)
 for USER in $LOCAL_USERS
 do
   if [[ ! ${LDAP_USERS[$USER]+_} ]]; then
+    echo "Removing authorized_keys file for ${USER}"
     rm -f /home/$USER/.ssh/authorized_keys
   fi
 done
